# docs/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentAuditResponse": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/IncidentAuditEvent"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "IncidentAuditResponse",
        "type": "object"
      },
      "IncidentBulkPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "incident_ids": {
            "items": {
              "type": "string"
            },
            "title": "Incident Ids",
            "type": "array"
          },
          "reopen": {
            "default": false,
            "title": "Reopen",
            "type": "boolean"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "open",
                  "acknowledged",
                  "escalated",
                  "closed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "user": {
            "default": "analyst_1",
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "incident_ids"
        ],
        "title": "IncidentBulkPatchRequest",
        "type": "object"
      },
      "IncidentBulkPatchResponse": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/LifecycleIncidentResponse"
            },
            "title": "Incidents",
            "type": "array"
          },
          "updated_count": {
            "title": "Updated Count",
            "type": "integer"
          }
        },
        "required": [
          "updated_count",
          "incidents",
          "errors"
        ],
        "title": "IncidentBulkPatchResponse",
        "type": "object"
      },
...
      "IncidentInvestigationResponse": {
        "additionalProperties": false,
        "properties": {
          "escalation_conditions": {
            "items": {
              "type": "string"
            },
            "title": "Escalation Conditions",
            "type": "array"
          },
          "false_positives": {
            "items": {
              "type": "string"
            },
            "title": "False Positives",
            "type": "array"
          },
          "pivot_queries": {
            "items": {
              "$ref": "#/components/schemas/SiemPivotQuery"
            },
            "title": "Pivot Queries",
            "type": "array"
          },
          "playbook_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playbook Id"
          },
          "playbook_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playbook Name"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "title": "Questions",
            "type": "array"
          },
          "triage": {
            "items": {
              "type": "string"
            },
            "title": "Triage",
            "type": "array"
          }
        },
        "required": [
          "questions",
          "pivot_queries",
          "false_positives",
          "escalation_conditions"
        ],
        "title": "IncidentInvestigationResponse",
        "type": "object"
      },
...
      "IncidentPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "reopen": {
            "default": false,
            "title": "Reopen",
            "type": "boolean"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "open",
                  "acknowledged",
                  "escalated",
                  "closed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "user": {
            "default": "analyst_1",
            "title": "User",
            "type": "string"
          }
        },
        "title": "IncidentPatchRequest",
        "type": "object"
      },
...
      "LifecycleIncidentResponse": {
        "properties": {
          "affected_entities": {
            "items": {
              "type": "string"
            },
            "title": "Affected Entities",
            "type": "array"
          },
          "ai_pipeline_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIPipelineMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "chain_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Confidence"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "correlation_flags": {
            "items": {
              "type": "string"
            },
            "title": "Correlation Flags",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "dest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest"
          },
          "detection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Id"
          },
          "effective_priority": {
            "title": "Effective Priority",
            "type": "string"
          },
          "evidence": {
            "$ref": "#/components/schemas/IncidentEvidenceNew"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "explanation": {
            "$ref": "#/components/schemas/IncidentExplanationNew"
          },
          "first_seen": {
            "title": "First Seen",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "investigation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentInvestigationNew"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_stale": {
            "title": "Is Stale",
            "type": "boolean"
          },
          "last_seen": {
            "title": "Last Seen",
            "type": "string"
          },
          "mitre": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentMitreMappingNew"
              },
              {
                "type": "null"
              }
            ]
          },
          "mitre_technique": {
            "default": "T1110",
            "title": "Mitre Technique",
            "type": "string"
          },
          "playbook_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playbook Id"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Level"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "sla_action_required": {
            "title": "Sla Action Required",
            "type": "boolean"
          },
          "sla_breached": {
            "title": "Sla Breached",
            "type": "boolean"
          },
          "sla_deadline_minutes": {
            "title": "Sla Deadline Minutes",
            "type": "integer"
          },
          "source": {
            "default": "auth",
            "enum": [
              "auth",
              "ai_pipeline"
            ],
            "title": "Source",
            "type": "string"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "status": {
            "default": "open",
            "enum": [
              "open",
              "acknowledged",
              "escalated",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "$ref": "#/components/schemas/IncidentSubjectNew"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "time_in_state_seconds": {
            "title": "Time In State Seconds",
            "type": "integer"
          },
          "time_since_alert_seconds": {
            "title": "Time Since Alert Seconds",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "valid_next_statuses": {
            "items": {
              "type": "string"
            },
            "title": "Valid Next Statuses",
            "type": "array"
          }
        },
        "required": [
          "incident_id",
          "type",
          "severity",
          "confidence",
          "first_seen",
          "last_seen",
          "affected_entities",
          "evidence_count",
          "source_count",
          "summary",
          "recommended_actions",
          "explanation",
          "subject",
          "evidence",
          "is_stale",
          "time_since_alert_seconds",
          "time_in_state_seconds",
          "sla_deadline_minutes",
          "sla_breached",
          "sla_action_required",
          "effective_priority",
          "valid_next_statuses"
        ],
        "title": "LifecycleIncidentResponse",
        "type": "object"
      },
...
      "PlaybookResponse": {
        "additionalProperties": false,
        "properties": {
          "chain": {
            "items": {
              "type": "string"
            },
            "title": "Chain",
            "type": "array"
          },
          "chain_context": {
            "default": "",
            "title": "Chain Context",
            "type": "string"
          },
          "chain_steps": {
            "items": {
              "type": "string"
            },
            "title": "Chain Steps",
            "type": "array"
          },
          "correlation": {
            "default": "",
            "title": "Correlation",
            "type": "string"
          },
          "deployment_context": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Deployment Context",
            "type": "object"
          },
          "detection_logic": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Detection Logic",
            "type": "object"
          },
          "detections": {
            "$ref": "#/components/schemas/PlaybookDetections"
          },
          "escalation": {
            "$ref": "#/components/schemas/PlaybookEscalation"
          },
          "escalation_criteria": {
            "items": {
              "type": "string"
            },
            "title": "Escalation Criteria",
            "type": "array"
          },
          "false_positive_logic": {
            "default": "",
            "title": "False Positive Logic",
            "type": "string"
          },
          "false_positives": {
            "items": {
              "type": "string"
            },
            "title": "False Positives",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "risk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaybookRisk"
              },
              {
                "type": "null"
              }
            ]
          },
          "technique_id": {
            "title": "Technique Id",
            "type": "string"
          },
          "triage": {
            "items": {
              "type": "string"
            },
            "title": "Triage",
            "type": "array"
          }
        },
        "required": [
          "id",
          "technique_id",
          "name",
          "detections",
          "triage",
          "false_positives",
          "escalation"
        ],
        "title": "PlaybookResponse",
        "type": "object"
      },
...
      "RunsIncidentsResponse": {
        "additionalProperties": false,
        "properties": {
          "incident_count": {
            "title": "Incident Count",
            "type": "integer"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/IncidentNew"
            },
            "title": "Incidents",
            "type": "array"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "required": [
          "incident_count",
          "incidents"
        ],
        "title": "RunsIncidentsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SecureWatch Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/incidents/bulk": {
      "patch": {
        "operationId": "bulk_patch_incidents_incidents_bulk_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentBulkPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentBulkPatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Patch Incidents",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleIncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      },
      "patch": {
        "operationId": "patch_incident_incidents__incident_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleIncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}/audit": {
      "get": {
        "operationId": "get_incident_audit_incidents__incident_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentAuditResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Audit",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}/investigation": {
      "get": {
        "operationId": "get_incident_investigation_incidents__incident_id__investigation_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentInvestigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Investigation",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}/playbook": {
      "get": {
        "operationId": "get_incident_playbook_incidents__incident_id__playbook_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Playbook",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}/secure-context": {
      "get": {
        "operationId": "get_secure_context_incidents__incident_id__secure_context_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user",
            "required": false,
            "schema": {
              "default": "analyst_1",
              "title": "User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Secure Context",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}/secure-context/audit": {
      "get": {
        "operationId": "get_secure_context_audit_incidents__incident_id__secure_context_audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Secure Context Audit",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}/spl": {
      "get": {
        "operationId": "get_incident_spl_incidents__incident_id__spl_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentSplResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident Spl",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/playbooks/incident/{incident_id}": {
      "get": {
        "operationId": "get_playbook_for_incident_playbooks_incident__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Playbook For Incident",
        "tags": [
          "playbooks"
        ]
      }
    },
...
    "/runs/{run_id}/incidents": {
      "get": {
        "description": "Get detected incidents for a specific run.",
        "operationId": "get_incidents_runs__run_id__incidents_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsIncidentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incidents",
        "tags": [
          "retrieval"
        ]
      }
    },
...
  }
}